Simplify refresh command path
